requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2>=2.8
drf-spectacular==0.28.0
pillow == 11.1.0